Complete feature engineering and target construction
